Pull state and snackbar up a level; pass to status update component too
