+++ moped-editor/src/views/dashboard/DashboardTimelineModal.js
@@ -1,4 +1,4 @@
-import React, { useCallback, useState } from "react";
+import React, { useState } from "react";
 import {
   Box,
   CardContent,
@@ -15,7 +15,6 @@ import { useQuery } from "@apollo/client";
 import ApolloErrorHandler from "src/components/ApolloErrorHandler";
 import ProjectMilestones from "../projects/projectView/ProjectMilestones";
 import ProjectPhases from "../projects/projectView/ProjectPhases";
-import FeedbackSnackbar from "src/components/FeedbackSnackbar";
 
 const useStyles = makeStyles((theme) => ({
   clickableDiv: {
@@ -28,6 +27,7 @@ const DashboardTimelineModal = ({
   projectId,
   projectName,
   dashboardRefetch,
+  handleSnackbar,
   children,
 }) => {
   const classes = useStyles();
@@ -49,38 +49,6 @@ const DashboardTimelineModal = ({
     setIsDialogOpen(false);
     dashboardRefetch();
   };
-
-  /* Snackbar state and handler for phase and milestone update feedback */
-  const [snackbarState, setSnackbarState] = useState(false);
-
-  /**
-   * Wrapper around snackbar state setter
-   * @param {boolean} open - The new state of open
-   * @param {String} message - The message for the snackbar
-   * @param {String} severity - The severity color of the snackbar
-   * @param {Object} error - The error to be displayed and logged
-   */
-  const handleSnackbar = useCallback(
-    (open, message, severity, error) => {
-      // if there is an error, render error message,
-      // otherwise, render success message
-      if (error) {
-        setSnackbarState({
-          open: open,
-          message: `${message}. Refresh the page to try again.`,
-          severity: severity,
-        });
-        console.error(error);
-      } else {
-        setSnackbarState({
-          open: open,
-          message: message,
-          severity: severity,
-        });
-      }
-    },
-    [setSnackbarState]
-  );
 
   return (
     <>
@@ -140,10 +108,6 @@ const DashboardTimelineModal = ({
           </ApolloErrorHandler>
         </DialogContent>
       </Dialog>
-      <FeedbackSnackbar
-        snackbarState={snackbarState}
-        handleSnackbar={handleSnackbar}
-      />
     </>
   );
 };
